added customer

+++ app/src/main/java/com/example/sufatura/domain/model/TariffBreakdown.kt
@@ -1,7 +1,9 @@
 package com.example.sufatura.domain.model
 
+import com.example.sufatura.data.local.entities.Tariff
+
 data class TariffBreakdown(
     val tariff: Tariff,      // Tarife bilgisi
     val consumption: Int,    // Bu dilimde harcanan tüketim miktarı
-    val cost: Double         // Bu dilimdeki maliyet
+    val cost: Int         // Bu dilimdeki maliyet
 )
